refactor(graphql): move index.ts's helper funcs & types to own files
move from graphql/src/index.ts:
- types to graphql/src/types.ts:
  - Venue
  - EsHitSource
  - EsHit
  - EsResults
  - GetCursor
- helper functions to graphql/src/utils.ts:
  - edgesFromEsResults
  - getHits
  - getTodayString
  - validateOrderByArguments

also:
 - add comments to the moved unspecific types so its clear that it'd be
   an improvement to somehow combine the TypeScript types and the
   GraphQL types

refs US-171 (trying to make the graphql/src/index.ts more manageable)
